% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{c_chmod}
\alias{c_chmod}
\title{Concrete class constructor}
\usage{
c_chmod(Y, X, modclass = "lm")
}
\arguments{
\item{Y}{vector of the outcome variable}

\item{X}{matrix of covariates; excepting intercepts 1's}

\item{modclass}{the class of model. Currently, "lm" for linear regression, "logit" (logit model),
"qreg" (quantile regression), "probit" (probit model), "gammainverse" (gamma with inverse link),
"gammalog" (gamma with log link), "poissonlog" (poisson model with log link),
"poissonidentity" (poisson with identity link), "poissonsqrt" (poisson with sqrt link),
"negbin" (negative binomial) are supported.}
}
\value{
object an object list with class attribute modclass.
}
\description{
A function for constructing functions for concrete classes of models for the \code{chmod()} family of
 of functions.
}
