% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CCRls.coord}
\alias{CCRls.coord}
\title{Linear regression via coordinate descent with covariate clsutering}
\usage{
CCRls.coord(Y, X, k, nC, ...)
}
\arguments{
\item{Y}{vector of outcome variable}

\item{X}{matrix of covariates. Should not include 1's for the intercept}

\item{k}{number of clusters}

\item{nC}{first nC-1 covariates in X not to cluster. Must be at least 1 for the intercept}

\item{...}{additional parameters to be passed to \link[stats]{lm}}
}
\value{
\code{betas}  parameter estimates (intercept first),

\code{iter}  number of iterations,

\code{dev}  increment in the objective function value at convergence

\code{fval} objective function value at convergence
}
\description{
This function is a wrapper for \code{linrclus}. It returns clustered parameters with
function values like residual sum of squares, and the criterion of choice (BIC, AIC)
}
\examples{
##Not run:
}
