% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ch.model}
\alias{ch.model}
\title{Choice of model}
\usage{
ch.model(Y, X, model = "lm", ...)
}
\arguments{
\item{Y}{the dependent or response variable}

\item{X}{design matrix (without intercept)}

\item{model}{a string for the desire model name. The following are supported
"lm", "logit", "probit", "gammainverse", "gammaidentity", "gammalog",
"poissonlog", "poissonsqrt", "negbin", "quantreg"}

\item{...}{additional model arguments}
}
\value{
mobj fitted model object
}
\description{
This function is used to choose th appropriate model
}
\examples{
## Not run: summary(ch.model(cars$speed,cars$dist,"gammaidentity"))
}
