% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{chmod.gammainverse}
\alias{chmod.gammainverse}
\title{Regression - gammainverse class}
\usage{
\method{chmod}{gammainverse}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and X - design matrix of class "probit"}

\item{...}{additional parameters to be passed to \code{\link[stats]{glm}}}
}
\value{
fitted model object
}
\description{
A gamma regression implementation for the "gammainverse" class. It uses \code{\link[stats]{glm}}
with the Gamma link function set to "inverse"
}
\examples{
chmod(c_chmod(Y=women$height,X=women$weight,modclass="gammainverse"))
}
