% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{chmod.logit}
\alias{chmod.logit}
\title{Regression - logit class}
\usage{
\method{chmod}{logit}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and Xmat - design matrix of class "logitcd"}

\item{...}{additional parameters to be passed to \code{\link[stats]{glm}}}
}
\value{
fitted model object
}
\description{
A logit regression implementation for the "logitcd" class. It uses \code{\link[stats]{glm}}
}
\examples{
chmod(c_chmod(Y=women$height<=50,X=women$weight,modclass="logit"))
}
