% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{chmod.negbin}
\alias{chmod.negbin}
\title{Regression - negbin class}
\usage{
\method{chmod}{negbin}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and X - design matrix of class "negbin"}

\item{...}{additional parameters to be passed to \code{\link[MASS]{glm.nb}}}
}
\value{
fitted model object
}
\description{
A negative binomial regression implementation for the "negbin" class. It uses \code{\link[MASS]{glm.nb}}
}
