% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{chmod.poissonidentity}
\alias{chmod.poissonidentity}
\title{Regression - poissonidentity class}
\usage{
\method{chmod}{poissonidentity}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and X - design matrix of class "poissonidentity"}

\item{...}{additional parameters to be passed to \code{\link[stats]{glm}}}
}
\value{
fitted model object
}
\description{
A poisson regression implementation for the "poissonidentity" class. It uses \code{\link[stats]{glm}}
with the poisson link function set to "identity"
}
\examples{
chmod(c_chmod(Y=women$height,X=women$weight,modclass="poissonidentity"))
}
