% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{chmod.poissonlog}
\alias{chmod.poissonlog}
\title{Regression - poissonlog class}
\usage{
\method{chmod}{poissonlog}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and X - design matrix of class "poissonlog"}

\item{...}{additional parameters to be passed to \code{\link[stats]{glm}}}
}
\value{
fitted model object
}
\description{
A poisson regression implementation for the "poissonlog" class. It uses \code{\link[stats]{glm}}
with the poisson link function set to "log"
}
\examples{
chmod(c_chmod(Y=women$height,X=women$weight,modclass="poissonlog"))
}
