% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{chmod.poissonsqrt}
\alias{chmod.poissonsqrt}
\title{Regression - poissonsqrt class}
\usage{
\method{chmod}{poissonsqrt}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and X - design matrix of class "poissonsqrt"}

\item{...}{additional parameters to be passed to \code{\link[stats]{glm}}}
}
\value{
fitted model object
}
\description{
A poisson regression implementation for the "poissonsqrt" class. It uses \code{\link[stats]{glm}}
with the poisson link function set to "sqrt"
}
\examples{
chmod(c_chmod(Y=women$height,X=women$weight,modclass="poissonsqrt"))
}
