% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{chmod.qreg}
\alias{chmod.qreg}
\title{Regression - qreg class}
\usage{
\method{chmod}{qreg}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and X - design matrix of class "qreg"}

\item{...}{additional parameters to be passed to \code{\link[quantreg]{rq}}, for example
\code{tau}}
}
\value{
fitted model object
}
\description{
A quantile regression implementation for the "qreg" class. It uses \code{\link[quantreg]{rq}}
}
\examples{
chmod(c_chmod(Y=women$height,X=women$weight,modclass="qreg"),tau=0.45)
}
