% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{clfun2}
\alias{clfun2}
\title{CCR at k}
\usage{
clfun2(Y, X, Xnc = NULL, clus, k, modclass = "lm", ...)
}
\arguments{
\item{Y}{vector of dependent variable Y}

\item{X}{CCR design matrix (without intercept) and covariates not clustered}

\item{Xnc}{matrix of covariates left out of CCR's clustering; defaults to \code{NULL} if all
covariates, except the intercept, are clustered.}

\item{clus}{vector of cluster assignments; must comprise positive integers only. Its sorted
unique values should number from 1 to k}

\item{k}{number of clusters - this is also the length of  \eqn{\delta}}

\item{modclass}{class of model for CCR estimation - see \link{c_chmod} for details.}

\item{...}{additonal parameters to be passed to the model, eg. tau for quantile regression.}
}
\value{
model object; may be glm, lm, or rq class of model depending on \code{modclass} chosen
}
\description{
Run lower-dimension regression given CCR objects \code{clus} and a specified \code{k}
}
