% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{linrclus}
\alias{linrclus}
\title{Linear regression with Conjugate Gradient Solver}
\usage{
linrclus(Y, X, k, coefs, clus, clusmns, nC = 1)
}
\arguments{
\item{Y}{vector of outcome variable}

\item{X}{matrix of covariates}

\item{k}{number of clusters}

\item{coefs}{vector of coefficients as starting values}

\item{clus}{vector of covariate cluster assignments as starting values}

\item{clusmns}{vector k cluster parameter centers}

\item{nC}{first nC-1 covariates in X not to cluster. Must be at least 1 for the intercept}
}
\value{
clus cluster assignments

coefs vector of coefficients as starting values

clusmns vector of cluster means
}
\description{
This function implements linear regression by solving the normal equations. It
constructs matrices X'X and X'Y and solves for beta in X'X*beta=X'Y
using the iterative Conjugate Gradient solver of system of linear equations.
}
\examples{
set.seed(14) #Generate data
N = 1000; (bets = rep(-2:2,4)); p = length(bets); X = matrix(rnorm(N*p),N,p)
Y = cbind(1,X)\%*\%matrix(c(0.5,bets),ncol = 1)
begin_v<- rep(NA,p)
for (j in 1:p) {
 begin_v[j] = coef(lm(Y~X[,j]))[2]
}
set.seed(12); klus_obj<- kmeans(begin_v,centers = 5)
linrclus(Y,cbind(1,X),k=5,coefs=c(0,begin_v),clus=klus_obj$cluster,clusmns=klus_obj$centers)
}
