% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{chmod.probit}
\alias{chmod.probit}
\title{Regression - probit class}
\usage{
\method{chmod}{probit}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and X - design matrix of class "probit"}

\item{...}{additional parameters to be passed to \code{\link[stats]{glm}}}
}
\value{
fitted model object
}
\description{
A probit regression implementation for the "probit" class. It uses \code{\link[stats]{glm}}
with the binomial link set to "probit"
}
\examples{
chmod(c_chmod(Y=women$height<=50,X=women$weight,modclass="probit"))
}
